:sparkles: feat: 인증된 유저가 CUD 작업을 할 수 있게 수정
- Post controller CUD 부분 인증을 위한 authentication 객체 주입
- Post service memberRepository 의존성 주입 및 CUD 메서드에 인증에 대한 로직 추가, 요청하는 Id와 db에 저장된 Id가 맞는지?
- Post dto 사용자의 식별 정보 삭제 'nickname'
- Post entity db에 fk 설정
- UserPrincipal member 로 바꿔주는 로직 추가
- Member entity Id를 생성자쪽으로 변경
- SecurityConfig CUD 작업에 대한 권한을 다시 수정

+++ src/main/kotlin/com/teamsparta/assignment/domain/post/service/PostService.kt
@@ -1,11 +1,14 @@
 package com.teamsparta.assignment.domain.post.service
 
 import com.teamsparta.assignment.domain.exception.ModelNotFoundException
+import com.teamsparta.assignment.domain.member.model.Member
+import com.teamsparta.assignment.domain.member.repository.MemberRepository
 import com.teamsparta.assignment.domain.post.dto.CreatePostRequest
 import com.teamsparta.assignment.domain.post.dto.PostResponse
 import com.teamsparta.assignment.domain.post.dto.UpdatePostRequest
 import com.teamsparta.assignment.domain.post.model.Post
 import com.teamsparta.assignment.domain.post.repository.PostRepository
+import com.teamsparta.assignment.infra.security.dto.UserPrincipal
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -13,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class PostService (
+    private val memberRepository: MemberRepository,
     private val postRepository : PostRepository
 ) {
 
@@ -32,9 +36,11 @@ class PostService (
 
     }
 
-    fun create(request: CreatePostRequest) : PostResponse {
+    fun create(request: CreatePostRequest, userPrincipal: UserPrincipal) : PostResponse {
 
-        val post = Post(request.nickname, request.title, request.description)
+        val member = memberRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException("User", userPrincipal.id)
+
+        val post = Post(request.title, request.description, member)
 
         val createdPost = postRepository.save(post)
 
@@ -43,16 +49,24 @@ class PostService (
     }
 
     @Transactional
-    fun update(postId: Long, request: UpdatePostRequest): PostResponse {
+    fun update(postId: Long, request: UpdatePostRequest, member: Member): PostResponse {
 
         val foundPost = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
 
+        foundPost.checkAuthorization(member)
+
         foundPost.updatePostField(request)
+
+        // 수정 요청 아이디와 foundPost 아이디가 같은지 확인
 
         return PostResponse.from(foundPost)
     }
 
-    fun delete(postId: Long) {
+    fun delete(postId: Long, userPrincipal: UserPrincipal) {
+
+        val foundPost = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+
+        foundPost.checkAuthorization(userPrincipal.to())
 
         postRepository.deleteById(postId)
